f moved chart to standalone component
