+++ frontend/src/components/TestChart/TestChart.module.css
@@ -1,0 +1,3 @@
+.chart {
+  padding-top: 2.5rem;
+}
